Remove the localStorage utils module
